+++ src/example/counter/utils.ts
@@ -1,6 +1,6 @@
-import { Resource } from "march-hare";
+import { app } from "../app.ts";
 import { User } from "./types";
 
-export const user = Resource(async (): Promise<User> => {
+export const user = app.Resource<User>(async () => {
   return { name: "Adam", age: 30 };
 });