+++ src/library/index.ts
@@ -1,4 +1,4 @@
-export { App, useApp } from "./app/index.tsx";
+export { App } from "./app/index.tsx";
 
 export { Action } from "./action/index.ts";
 export { Distribution, Lifecycle } from "./types/index.ts";
@@ -6,7 +6,6 @@ export { With } from "./with/index.ts";
 
 export { Boundary } from "./boundary/index.tsx";
 
-export { Resource } from "./resource/index.ts";
 export { Cache } from "./cache/index.ts";
 
 export { Reason, Aborted } from "./error/index.ts";
@@ -15,6 +14,8 @@ export { annotate } from "./annotate/index.ts";
 export { Operation, Op, State } from "immertation";
 
 export * as utils from "./utils/index.ts";
+
+export * as shared from "./shared/index.ts";
 
 export type { Fault } from "./error/index.ts";
 export type { Adapter } from "./cache/index.ts";